Add control API with app version management
