+++ README.md
@@ -30,31 +30,29 @@ database.
 
 ## Swagger
 
-Open `http://localhost:5067/swagger` for interactive documentation.
-Use the `POST /api/notifications/form` operation to send a notification with
+Open `http://localhost:5067/swagger` for interactive documentation. Use the
+`POST /api/Control/send-notification` operation to send a notification with
 `multipart/form-data` (including an optional `file` field). To send a JSON payload
-with Base64 data, call `POST /api/notifications`. `GET /api/notifications` lists the
-latest notifications (the server retains only the most recent 20).
-
-Device endpoints:
-
-- `POST /api/devices/{id}/version` – report a device's current app version and optional card code.
-- `GET /api/devices/{id}/notifications` – fetch notifications the device hasn't
-  received yet.
-- `POST /api/devices/{id}/notifications/{notificationId}/read` – mark a notification as read; updates `LastSeen` and returns the device's card code and version.
+with Base64 data, call `POST /api/Control/send-notification-json`. `GET
+/api/Control/get-notifications` lists the latest notifications (the server
+retains only the most recent 20).
 
 ## API Endpoints
 
+| # | API | Chức năng chính |
+| - | --- | ---------------- |
+| 1 | `POST /api/Control/send-notification-json` | Gửi thông báo dạng JSON (đính kèm file qua base64). |
+| 2 | `POST /api/Control/send-notification` | Gửi thông báo với file upload dạng `multipart/form-data`. |
+| 3 | `GET /api/Control/get-notifications` | Lấy danh sách thông báo hiện có (có phân trang). |
+| 4 | `POST /api/Control/clear-notifications` | Xóa toàn bộ thông báo khỏi hàng đợi. |
+| 5 | `GET /api/Control/notifications-stream` | Stream SSE để nhận thông báo thời gian thực. |
+| 6 | `GET /api/Control/app-version` | Trả về thông tin phiên bản ứng dụng mới nhất. |
+| 7 | `POST /api/Control/app-version/upload` | Upload bản build mới (APK/IPA) và cập nhật manifest. |
+| 8 | `GET /api/Control/app-version/download` | Tải xuống bản build mới cho thiết bị (Android/iOS). |
+| 9 | `HEAD /api/Control/app-version/download` | Kiểm tra metadata (kích thước, SHA256, content type) của bản build mới. |
+
 | Method | Path | Description |
 | ------ | ---- | ----------- |
-| `POST` | `/api/notifications` | Send a notification as JSON with optional Base64 file data and `targetVersion`. |
-| `POST` | `/api/notifications/form` | Send a notification via multipart form-data with optional `targetVersion` and file. |
-| `GET` | `/api/notifications` | List stored notifications (most recent first, capped at 20). |
-| `POST` | `/api/notifications/clear` | Remove all stored notifications and delivery records. |
-| `GET` | `/api/notifications/stream` | Stream new notifications in real time using Server-Sent Events. |
-| `POST` | `/api/devices/{id}/version` | Report the current version of a device along with optional `cardCode`. |
-| `GET` | `/api/devices/{id}/notifications` | Fetch notifications a device has not yet received. |
-| `POST` | `/api/devices/{id}/notifications/{notificationId}/read` | Mark a notification as read and update the device's last-seen timestamp. |
 | `GET` | `/uploads/{filename}` | Download an uploaded file (URL returned in notification). |
 
 ## Postman / cURL
@@ -62,13 +60,13 @@ Device endpoints:
 Send a multipart request (form-data) with fields `id`, `title`, `body`, optional `link`, optional `targetVersion`, and an optional `file`:
 
 ```
-POST http://localhost:5067/api/notifications/form
+POST http://localhost:5067/api/Control/send-notification
 ```
 
 Or send JSON when you already have the file as Base64:
 
 ```
-POST http://localhost:5067/api/notifications
+POST http://localhost:5067/api/Control/send-notification-json
 Content-Type: application/json
 
 {
@@ -85,24 +83,10 @@ Content-Type: application/json
 Retrieve messages:
 
 ```
-GET http://localhost:5067/api/notifications
+GET http://localhost:5067/api/Control/get-notifications
 ```
 
 Responses include a `fileUrl` you can open to download or preview the uploaded file.
-
-Report device details:
-
-```
-POST http://localhost:5067/api/devices/device123/version
-Content-Type: application/json
-
-{
-  "version": "1.0.0",
-  "cardCode": "CARD123"
-}
-
-POST http://localhost:5067/api/devices/device123/notifications/{notificationId}/read
-```
 
 SQL Server DDL for the backing tables is available in `sql/create_tables.sql`.
